Browser: Screen will garbage after rotate the phone.
This is Google's original bug and can also be reproduced on
Samsung Nexus phone(4.3).

The current phenomenon has a tremendous impact on User
Experience, but it's difficult to get the root cause quickly.
Therefore, for avoiding this issue temporarily, let Browser
wait about 300ms and then refresh WebView in onConfigurationChanged
method when rotate DUT.

CRs-Fixed: 520353

Change-Id: I4e4b0aaea6e7f306eda1b31069f5b737ef76e030