Browser: cut clear state funciton from onDestroy
It will cause a problem that when we change system language then
back to browser,the browser will goto homepage instead of staying
at the current page because the browser activity will be recreated.
So there is no need to add clear state into onDestory.The state will
be cleared when user press back key or "exit" menu to exit the browser.

CRs-Fixed: 530324

Change-Id: I52c1063ec55c4265c5b2632bbcf1fa3f808ac0ca