Browser: Add tips for new message received in full-screen mode
When the browser is running in full-screen mode,the status bar is
hidden,so the new message received notification can't be viewed
timely by the user.

Add a receiver to monitor the custom broadcast from MMS, and using
toast to alert users the new message received in full-screen mode.

CRs-Fixed: 534580

Change-Id: Ie3f3d6896ec30caa07303fe6e8917676e0e629a0