Browser: fix can not search with browser
If the current search engine is not the default engine,
it will finish the current activity but not start a new
activity, this make the user can not search with browser.

If the search engine is a instance of DefaultSearchEngine,
we finish the current activity, otherwise, the activity
should not be finished.

CRs-Fixed: 535416

Change-Id: Ieb323dafea834e47d1c35c1dff929d638c47f31d