Browser: fix can not show most_visited page
The original documentation for UriMatcher class indicated
a leading slash was possible but the code did not support it.
So google delete the leading slash.But the slash will be deleted
even if the path is "/".So UriMatcher can't match the uri whose
path is "/".

Add a string to make uri match successfully.

CRs-Fixed: 522453

Change-Id: I2b4a8e517aad7bf166575b22e370461976b11199