Fix Url text not getting deleted on pressing delete key.
Fixed issue with url is not deleted completely when clicking the
address bar and delete text by the "delete" key in the andoid
keyboard.

This issue is specific to Google keyboard on Kitkat and it seems
that keyboard is not sending 'KEYCODE_DEL' event on pressing
backspace/delete on soft keyboard.

Change-Id: I86a91897ea45dc48e4a005d51800ef241ff1073a