Fixed rendering of noreferrer links in new tab
Rendering of a link that opens in a new tab and the
link contains "rel=noreferrer" tag was failing.
The noreferrer tag causes opener to be suppressed.
Due to a race condition in the code when OpenURL is
called the web content delegate is not yet configured.

This fix informs browser that opener is suppressed.
The browser calls explicit loadURL for this case.

Change-Id: Id9f45a934a5fe7df1f06dc753f93053556943249