Fix remaining createScreenshot references
Fixed references to createScreenshot method which was returning a null
bitmap. createScreenshot method was implemented using the older
getViewportBitmap synchronous API which was deprecated. WebView exposes
an asynchronous method getContentBitmapAsync to retrieve a bitmap of the
current view. Modified calls to use the getContentBitmapAsync API.
This also fixes the issue seen where an async task was fired in a loop
to retrieve the bitmap resulting in wasted CPU cycles.

Change-Id: Ic6d8385d41d90c58fdffcebcbca7ebd68381530f