Clean visual style: move to Light, add Material support, change Drawables (progress, icons).
This code alters the visual style moving the APP to a Light theme.
Parents of the theme are either Theme.Holo.Light or Theme.Material.Light
depending on which API we start the APK from. Needs SDK 21 to build.

Other than the visual changes, this commit changes:
 - name of the APP to 'Browser'
 - animations and sequencing of the Tab switcher in {nav/anim}_screen
 - removal of the warning icon from the Exit dialog
 - change of the favicon appearance (removal of the double border)

Vast amount of unused resources are removed from the APK. Just XHDPI
drawables have been synthesized for now, and the Primary color was
used for the icons and is accessible in R.color.Primary.

Change-Id: If75cc051c5d4015383e96066cdb6507484e625d8