Update the orientation of the NavScreen on orientationChange.
Before, it was destroying all the UI objects and recreating them,
no matter if the tab switcher was visible or not. The new behavior
matches the old one visually, but happens in a fraction of the
time.

Also, this updates a bit the visibility of properties in the class.
This class should require more cleanup in the future.

Change-Id: I4e5be461aa954faf8fc6b8d50fa8dd865a72e564