Fix for blank titlebar after layout
Currently android DisplayList calculates the transparent
regions of layers on layout, taking translation of
views also into account. We have a bug in browser where
TitleBar becomes blank when it is translated out of viewport
and relayout happens due to orientation change.
The fix will translate the titlebar Y to 0 during layout
and translate it back to its original position after the
layout.

Change-Id: I8d66148bd164c4cdfd7b48f98f420174dc18fbd9