Added more details to crash log report
Exception handler gets a nested throwable as an argument. The nested
throwable can have multiple entries if the application handled the
exception and re-threw it (as the same or a different exception).

With this fix, exception handler will include information for all the
entries in the crash report. Additionally, it'll include browser's
readable version and build date.

Change-Id: I3e90ef073fbc50efc5ea3f5e486c2f27c033256b