Add vertical scrolling for displaying URL.
When viewing a long url in by performing
a long press in the Bookmark View.It used to get
trimmed out at 80 characters.

This change adds a vertical scrolling so that
the first 3 lines each 40 characters long are being shown to
the user and rest of them can viewed by scrolling up.

Change-Id: Ide8c5d2595d11d7b72e605319a07e2bb861937bf