Fix for Context Menu javascript url with image link
This fix relates to situations in which an image is embedded
in an anchor tag whose href is javascript schemed url.
Thus the long press performed on such urls returns an
UNKNOWN_TYPE which is not handled by Browser and
context menu is not shown to the user.

This change fixes the use case.

Change-Id: Ifce1043244163e56966eed3a64522068680366e2