MDM Do-Not-Track Restriction
Implements handling of MDM enforced restriction preventing the user
from updating the 'Do Not Track' checkbox in the Privacy & Security
menu.

When enforced by MDM, the 'Do Not Track' menu entry is greyed out, and
if tapped, presents a toast message informing the user that the
setting is managed by the administrator.  The checkbox reflects
the value set by MDM.

Included in this commit is the basic unit tests and a new
'MdmCheckBoxPreference' class that manages the display of the toast
message and makes it possible to update the state of the entry
immediately (instead of having to restart the menu to have a new
MDM setting take effect.

Change-Id: I931d80ccf80474120dba2d768a7a257e337bece2