MDM Incognito Mode Restriction
Implements handling of MDM enforced restriction preventing the user
from opening an incognito tab.

When enforced by MDM, the Incognito button on the tabs screen
is greyed out, and if tapped, presents a toast message informing
the user that the setting is managed by the administrator.

This commit includes basic unit tests.

Change-Id: I26fcf10cc954f63c8e4489dbac9ff148e90b0c94