MDM Third Party Cookies Restriction
Implemented Third Party Cookies (TPC) Restriction.
The control point is in the native code (cookie_settings.cc),
and we make use of the PrefServiceBridge to activate and
configure the restriction. To support this we've added two
new boolean preference keys (kMdmBlockThirdPartyCookiesManaged and
kMdmBlockThirdPartyCookies).

Note that when the restriction is in effect, it overrides the
'normal' preference setting (kBlockThirdPartyCookies), which
is present in SWE, but has no current UI element. If a UI is
eventually implemented to expose this setting, the MDM restriction
code will need to be updated to disable the UI feature when the
restriction is active.

Change-Id: I929c48b3a5f932daf63468baa7f327d3ce5c78be