MDM Managed Bookmarks
When passed a JSON description of a managed bookmarks tree through
the MDM mechanism, SWE populates a special 'Managed' bookmark
folder which builds the specified tree beneath it. If the JSON
managed bookmarks descriptor is the same as the currently enforced
descriptor, the current managed tree is left alone, otherwise the
current managed bookmarks tree is removed and the new descriptor
is used to populate a new Managed tree. If an MDM bundle is
received without a JSON managed bookmarks descriptior, any
existing 'Managed' bookmark tree is removed.

The bookmark elements in the 'Managed' tree are not editable
by the user (the edit and delete pop-up menu items are greyed-out),
and are rendered with a briefcase icon overlaid.

This commit adds the following new icons (new art created
in-house):
   res/drawable-xxhdpi/img_deco_mdm_badge_bright.png
   res/drawable-xxhdpi/img_deco_mdm_badge_dark.png

Other new files are the core implementation of the
ManagedBookmarkRestriction and the tests for it:
    ManagedBookmarksRestriction.java
    ManagedBookmarksRestrictionsTest.java

Also included in this commit are some debugging refinements
which required minor edits of unrelated MDM restrictions.

Change-Id: Iac7ae6bee788418d3d444577951a8d29af2a67f2