MDM Auto Fill Restriction
Implementation of MDM form auto fill restriction. When enforced,
the auto-fill menu item is greyed and unchecked, and if clicked,
results in the "Managed by your administrator" Toast message.

Disabling the restriction checks the menu item checkbox and ungreys
the menuitem, restoring full control of this feature to the user.

This patch also includes minor refactoring of DoNotTrackRestriction.

Change-Id: I2f36ea9263a70f0a9470900c3abc48d9376a4604