MDM Download Restriction support
This commit adds MDM management of the following if the restriction
is enabled via the MDM mechanism:
    1. The ability to download.
    2. The download directory under '/Phone Storage/', if
       downloading is permitted.

These restrictions are only effective on SWE provisioned in the
'Work' profile. Non-work-profile SWE will not be affected.

When the download capability has been restricted, clicking a download
link results in a Toast message (Managed by your administrator). A long-press
of the link still brings up a context menu, but the 'Save link as' menuitem
is disabled in this case.

If MDM sets the download directory, it is not setting the full path, just
the directory under '/Phone Storage'. The directory specified must include
a leading '/' i.e. '/NewDownloadDir'.

Change-Id: I13a44892a36925d6324b87f7655d79fb0776f8a9