Fix system Navigation bar covering menu when fullscreen
Fixed one case of the navigation bar covering the app
menu popup.

Needed to make the popup non-modal before showing it
and modal immediately after so that it can be
interacted with.

This will happen only when Immersive mode is in use
since testing doesn't work with non-modal popups.

CR-Fixed: 903500

Change-Id: Ia265d434a96822e8c1095e89aa42f67a335a13ff